CLN: underBar animation을 FilterView에서 처리하도록 변경
