+++ TwitterProject(Rx+MVVM)/Sources/View/ProfileViewController.swift
@@ -24,13 +24,11 @@ class ProfileViewController: UIViewController, ViewModelBindable {
     }
     override func viewWillAppear(_ animated: Bool) {
         navigationController?.navigationBar.isHidden = true
-        
     }
     // MARK: - Selectros
     // MARK: - Methods
     
     func configureUI() {
-        view.backgroundColor = .systemBackground
         collectionView.contentInsetAdjustmentBehavior = .never
         collectionView.register(TweetCell.self, forCellWithReuseIdentifier: reuseIdentifier)
         collectionView.register(ProfileHeader.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: headerIdentifier)
@@ -71,7 +69,8 @@ extension ProfileViewController: UICollectionViewDelegateFlowLayout {
     }
     func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
         let header = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: headerIdentifier, for: indexPath) as! ProfileHeader
-        
+//        header.backButton.rx.tap
+//            .bind(to: viewModel.input)
         return header
     }
 }
